Added support for multi-prupose playlists and some cleanup.
This update adds support for multi-purpose playlists. For example, if you have a playlist that includes Youtube URLs, Vimeo URLs and local files on your HDD.

Instead of changing `ytdl-format` once for all the playlist, now the script will do a one by one check, for more accuracy.

In the example mentioned above, the `ytdl-format` will only change for Youtube, not for Vimeo and script will not run if it's a local file, since it's not a URL stream.

The `ytdlLow` value will be used for matched domains in VSTREAM, otherwise `ytdlBest` will be used. So in essence, `ytdlBest` acts as the default `ytdl-format` you would normally set in `mpv.conf`